+++ R/herbivore_eat.R
@@ -2,9 +2,9 @@
 # plant and herbivore state
 herbivore_eat <- function(herbivore, plants) {
 
-  plant <- plants %>% filter(plant_id == herbivore$selected_plant_id)
+  plant <- plants %>% dplyr::filter(plant_id == herbivore$selected_plant_id)
 
-  prop_defence <- plant$b_def / plant$ms
+  prop_defence <- plant$bdef / plant$ms
   intake_rate <- (1 / herbivore$handling_time) / 1000 * (1 - prop_defence)
 
   # Calculate maximum possible intake given constraints
@@ -18,7 +18,7 @@ herbivore_eat <- function(herbivore, plants) {
   total_parts <- plant$ms
   leaf_intake <- potential_intake * (plant$ms / total_parts)
   stem_intake <- 0  # placeholder; modify if stems are separately defined
-  def_intake  <- potential_intake * (plant$b_def / total_parts)
+  def_intake  <- potential_intake * (plant$bdef / total_parts)
 
   # Update digestion vectors (first hour position)
   herbivore$digestion$bleaf[1]   <- herbivore$digestion$bleaf[1] + leaf_intake
@@ -27,9 +27,9 @@ herbivore_eat <- function(herbivore, plants) {
 
   # Update plant biomass after consumption
   plants <- plants %>% 
-    mutate(
+    dplyr::mutate(
       ms = ifelse(plant_id == herbivore$selected_plant_id, ms - potential_intake, ms),
-      b_def = ifelse(plant_id == herbivore$selected_plant_id, b_def - def_intake, b_def)
+      bdef = ifelse(plant_id == herbivore$selected_plant_id, bdef - def_intake, bdef)
     )
 
   # Update herbivore gut content